clean up a bit in network code
